Move proxy control into an abstract base class
